requirements.txt
-i https://pypi.org/simple
cffi==1.13.2
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
pycparser==2.19
pyparsing==2.4.5
pysoundfile==0.9.0.post1
python-dateutil==2.8.1
scipy==1.3.3
six==1.13.0
